userservice/requirements.txt
coreapi
coreschema
Django==5.1.6
djangorestframework
djangorestframework-simplejwt
django-smart-selects
python-decouple

